requirements.txt
opencv-python-headless
numpy
scikit-learn
pygame
PyYAML
